collect remaining td64index tests
